refactor(perf): reorganize benchmarking code into unified benchmarks/ directory
Consolidate scattered performance benchmarking code into a single, well-organized directory:

Structure:
- benchmarks/data/ - Measurement data (renamed from __report__/perf/*_raw)
  - prototype/ (was prototype_raw)
  - v0.2.0/ (was v0.2.0_raw)
- benchmarks/*.py - All benchmarking scripts in one place

Moved files:
- tests/performance/test_overhead.py → benchmarks/overhead_measurement.py
- tests/performance/workloads.py → benchmarks/workloads.py
- tests/performance/statistics_utils.py → benchmarks/statistics_utils.py
- scripts/compare_baselines.py → benchmarks/compare_baselines.py

Updated imports:
- All moved files now import from benchmarks.* namespace
- tests/performance/test_regression.py updated to import from benchmarks.*
- benchmarks/cprofile_comparison.py updated to use new import paths

Added documentation:
- benchmarks/README.md with quick start guide and methodology

Benefits:
- Single source of truth for all performance benchmarking
- Clearer separation from pytest tests (tests/performance/ now only has regression tests)
- Easier to find and maintain benchmarking code
- Better organization for stakeholder review
