+++ benchmarks/data/prototype/overhead_decorator_large_x10.json
