+++ benchmarks/data/prototype/overhead_decorator_medium_x1.json
