+++ benchmarks/data/prototype/overhead_decorator_medium_x10.json
