+++ benchmarks/data/prototype/overhead_decorator_medium_x100.json
