+++ benchmarks/data/prototype/overhead_decorator_small_x10.json
